# Conflicts:
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.idea/workspace.xml

+++ app/src/main/java/com/trekkstay/hotel/feature/reservation/domain/entities/reservation_room.kt
@@ -6,8 +6,8 @@ import com.trekkstay.hotel.feature.hotel.domain.entities.Media
 data class ReservationRoom(
     val hotelId:String,
     val type:String,
-    val originalPrice:Double,
-    val bookingPrice: Double,
+    val originalPrice:Int,
+    val bookingPrice: Int,
     val images: Media,
 
     )
